Bug 706986: Unexpected argument types should cause reflow. (r=dvander,mjrosenb)
UltraBlame original commit: 90e4885eada06554f09dc0cf204b4e3025d0db4e

+++ js/src/assembler/assembler/ARMAssembler.cpp
@@ -316,7 +316,7 @@ void ARMAssembler::dataTransferN(bool isLoad, bool isSigned, int size, RegisterI
     bool posOffset = true;
 
     
-    if (offset == 0x80000000) {
+    if (offset == int32_t(0x80000000)) {
         
         
         moveImm(offset, ARMRegisters::S0);
